requirements.txt

pyserial
pytz
